+++ clients/snowflake/dialect/ddl.go
@@ -1,9 +1,25 @@
 package dialect
 
 import (
+	"fmt"
+	"strings"
+
 	"github.com/artie-labs/transfer/lib/config/constants"
 	"github.com/artie-labs/transfer/lib/sql"
 )
+
+func (SnowflakeDialect) BuildCreateTableQuery(tableID sql.TableIdentifier, temporary bool, colSQLParts []string) string {
+	query := fmt.Sprintf("CREATE TABLE IF NOT EXISTS %s (%s)", tableID.FullyQualifiedName(), strings.Join(colSQLParts, ","))
+
+	if temporary {
+		// TEMPORARY Table syntax - https://docs.snowflake.com/en/sql-reference/sql/create-table
+		// PURGE syntax - https://docs.snowflake.com/en/sql-reference/sql/copy-into-table#purging-files-after-loading
+		// FIELD_OPTIONALLY_ENCLOSED_BY - is needed because CSV will try to escape any values that have `"`
+		return query + ` STAGE_COPY_OPTIONS = ( PURGE = TRUE ) STAGE_FILE_FORMAT = ( TYPE = 'csv' FIELD_DELIMITER= '\t' FIELD_OPTIONALLY_ENCLOSED_BY='"' NULL_IF='\\N' EMPTY_FIELD_AS_NULL=FALSE)`
+	} else {
+		return query
+	}
+}
 
 func (sd SnowflakeDialect) BuildAddColumnQuery(tableID sql.TableIdentifier, sqlPart string) string {
 	return sd.buildAlterColumnQuery(tableID, constants.Add, sqlPart)
@@ -11,4 +27,12 @@ func (sd SnowflakeDialect) BuildAddColumnQuery(tableID sql.TableIdentifier, sqlP
 
 func (sd SnowflakeDialect) BuildDropColumnQuery(tableID sql.TableIdentifier, colName string) string {
 	return sd.buildAlterColumnQuery(tableID, constants.Delete, colName)
+}
+
+func (SnowflakeDialect) BuildDescribeTableQuery(tableID sql.TableIdentifier) (string, []any, error) {
+	return fmt.Sprintf("DESC TABLE %s", tableID.FullyQualifiedName()), nil, nil
+}
+
+func (SnowflakeDialect) buildAlterColumnQuery(tableID sql.TableIdentifier, columnOp constants.ColumnOperation, colSQLPart string) string {
+	return fmt.Sprintf("ALTER TABLE %s %s COLUMN %s", tableID.FullyQualifiedName(), columnOp, colSQLPart)
 }
